requirements.txt
alembic==1.4.3
astropy==4.0.1.post1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
GeoAlchemy2==0.8.4
healpy==1.14.0
idna==2.10
infinity==1.5
intervals==0.9.0
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.19.2
orderedmultidict==1.0.1
Pillow==7.2.0
psycopg2-binary==2.8.6
pygcn==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
urllib3==1.25.10
validators==0.18.1
voevent-parse==1.0.3
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Components==0.10.4
